archive_pipeline/requirements.txt
pytest
pytest-cov
pylint
psycopg2-binary
python-dotenv